# docs/openapi.yaml
openapi: 3.0.3
info:
  title: 명장쇼핑몰 API
  description: 클라이언트 개발용 REST API 명세
  version: 1.0.0
servers:
  - url: http://localhost:3300
    description: 로컬 개발
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProductTag:
      type: string
      enum: [BEST, NEW, SALE]
...
paths:
...
  /products:
    get:
      tags: [Products]
      summary: 상품 목록 (페이지·정렬·필터 지원)
      description: 페이지(page) 기반 페이지네이션, sortBy/sortOrder로 정렬
      parameters:
        - name: brandId
          in: query
          schema: { type: string }
        - name: categoryId
          in: query
          schema: { type: string }
        - name: tag
          in: query
          schema: { $ref: '#/components/schemas/ProductTag' }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
        - name: minRating
          in: query
          schema: { type: number }
        - name: sortBy
          in: query
          schema: { type: string, enum: [price, createdAt, rating, salesCount, viewCount] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [ASC, DESC] }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer }
        - name: take
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: OK
    post:
      tags: [Products]
      summary: 상품 생성
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, price, brandId, categoryId]
              properties:
                name: { type: string }
                description: { type: string }
                htmlDescription: { type: string }
                price: { type: number, minimum: 0 }
                image: { type: string, format: uri }
                tag: { $ref: '#/components/schemas/ProductTag' }
                brandId: { type: string }
                categoryId: { type: string }
                stock: { type: integer, minimum: 0 }
                discountRate: { type: number, minimum: 0, maximum: 100 }
                rating: { type: number, minimum: 0, maximum: 5 }
                isActive: { type: boolean }
      responses:
        '201':
          description: Created
  /products/top:
    get:
      tags: [Products]
      summary: 인기 상품
      parameters:
        - name: limit
          in: query
          schema: { type: integer }
        - name: sortBy
          in: query
          schema: { type: string, enum: [price, createdAt, rating, salesCount, viewCount] }
      responses:
        '200':
          description: OK
  /products/brand/{brandId}:
    get:
      tags: [Products]
      summary: 브랜드별 상품
      description: page, take, sortBy, sortOrder 등 상품 목록과 동일 쿼리 지원
      parameters:
        - name: brandId
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: take
          in: query
          schema: { type: integer }
        - name: sortBy
          in: query
          schema: { type: string, enum: [price, createdAt, rating, salesCount, viewCount] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [ASC, DESC] }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
      responses:
        '200':
          description: OK
  /products/category/{categoryId}:
    get:
      tags: [Products]
      summary: 카테고리별 상품
      description: page, take, sortBy, sortOrder 등 상품 목록과 동일 쿼리 지원
      parameters:
        - name: categoryId
          in: path
          required: true
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: take
          in: query
          schema: { type: integer }
        - name: sortBy
          in: query
          schema: { type: string, enum: [price, createdAt, rating, salesCount, viewCount] }
        - name: sortOrder
          in: query
          schema: { type: string, enum: [ASC, DESC] }
        - name: search
          in: query
          schema: { type: string }
        - name: minPrice
          in: query
          schema: { type: number }
        - name: maxPrice
          in: query
          schema: { type: number }
      responses:
        '200':
          description: OK
  /products/{id}:
    get:
      tags: [Products]
      summary: 상품 상세
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
    patch:
      tags: [Products]
      summary: 상품 수정
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                htmlDescription: { type: string }
                price: { type: number }
                image: { type: string }
                tag: { $ref: '#/components/schemas/ProductTag' }
                brandId: { type: string }
                categoryId: { type: string }
                stock: { type: integer }
                discountRate: { type: number }
                rating: { type: number }
                isActive: { type: boolean }
      responses:
        '200':
          description: OK
    delete:
      tags: [Products]
      summary: 상품 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
...
  /wishlist:
    get:
      tags: [Wishlist]
      summary: 위시리스트 조회
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
    post:
      tags: [Wishlist]
      summary: 위시리스트 추가
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: string, format: uuid }
      responses:
        '201':
          description: Created
  /wishlist/{itemId}:
    delete:
      tags: [Wishlist]
      summary: 위시리스트 항목 삭제
      security: [{ bearerAuth: [] }]
      parameters:
        - name: itemId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
  /wishlist/product/{productId}:
    delete:
      tags: [Wishlist]
      summary: 위시리스트 상품별 삭제
      security: [{ bearerAuth: [] }]
      parameters:
        - name: productId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: No Content
  /simulator/sets:
    get:
      tags: [Simulator]
      summary: 시뮬레이터 세트 목록
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: isActive
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: integer }
        - name: take
          in: query
          schema: { type: integer }
      responses:
        '200':
          description: OK
    post:
      tags: [Simulator]
      summary: 시뮬레이터 세트 생성
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name, items]
              properties:
                name: { type: string, maxLength: 100 }
                description: { type: string }
                isActive: { type: boolean }
                items:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [productId, categoryId]
                    properties:
                      productId: { type: string }
                      categoryId: { type: string }
      responses:
        '201':
          description: Created
  /simulator/sets/{id}:
    get:
      tags: [Simulator]
      summary: 시뮬레이터 세트 단건 조회
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
    patch:
      tags: [Simulator]
      summary: 시뮬레이터 세트 수정
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                description: { type: string }
                isActive: { type: boolean }
                items:
                  type: array
                  items:
                    type: object
                    required: [productId, categoryId]
                    properties:
                      productId: { type: string }
                      categoryId: { type: string }
      responses:
        '200':
          description: OK
    delete:
      tags: [Simulator]
      summary: 시뮬레이터 세트 삭제
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '204':
          description: No Content
...
  /usability-services:
    post:
      tags: [Usability]
      summary: 사용성 서비스 신청
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, content]
              properties:
                title: { type: string, maxLength: 150 }
                content: { type: string }
      responses:
        '201':
          description: Created
  /usability-services/me:
    get:
      tags: [Usability]
      summary: 내 사용성 서비스 목록
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: OK
...